support nlargest & nsmallest
